Modify foreign key constraint for `UTxOToken` to refer to `Checkpoint`.
Ideally, we'd like to define a foreign key constraint on the UTxO table,
as each row in UTxOToken refers to a unique row in UTxO.

Unfortunately, there's a bug in our version of persistent that breaks
foreign key constraints where the foreign table has a manually-defined
table name. In the case of the UTxO, the table name is "utxo", and not
"u_tx_o" (which is what the auto-generated table name would be).

To work around this bug, we instead define a foreign key constraint on
the Checkpoint table. Under the current design of rollback, this gives
equivalent behaviour, as we always delete UTxO entries by deleting their
parent checkpoints.